Added Concurrency
- Previously had to process over 170 pages in a single thread
- Added 25 threads and cut the time by 4
