+++ pepperList.json
@@ -1,4 +1,4 @@
-[ {
+[ null, {
   "name" : "Jalapeno Purple",
   "altNames" : "Jalapeno Purple",
   "species" : "Capsicum annuum",
