+++ src/main/java/com/connellboyce/scraper/ScraperApplication.java
@@ -5,6 +5,9 @@ import com.connellboyce.scraper.scrapers.PepperScraper;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+
 @SpringBootApplication
 public class ScraperApplication {
